Fixed Minor Bugs in the teacherService and all functionalities work for the teacher. Now to focus on student and the respective services.
Some methods are to be migrated to student services due to the change in the model structure.

Minor bugs fixed :

NumberFormatException while deleting a question , can be triggered by inputting a blank line.

Save data method was not called when big changes occured in some methods in the view. (Fixed issues.)